requirements.txt
pandas
mistralai
python-dotenv
streamlit
matplotlib
seaborn